fix(about-page): move all client logic into separate components
