+++ front_end/src/app/(main)/questions/create/group/page.tsx
@@ -34,11 +34,7 @@ const GroupQuestionCreator: React.FC<{ searchParams: SearchParams }> = async ({
 
   return (
     <>
-      {community ? (
-        <CommunityHeader community={community} />
-      ) : (
-        <Header />
-      )}
+      {community ? <CommunityHeader community={community} /> : <Header />}
       <GroupForm
         // @ts-ignore
         subtype={
